requirements.txt
asttokens==2.0.8
backcall==0.2.0
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
decorator==5.1.1
easyocr==1.6.0
executing==1.0.0
fonttools==4.37.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
idna==3.3
imageio==2.21.2
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.8.6
ninja==1.10.2.3
numpy==1.23.2
# opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post3
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
ratelim==0.1.6
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
seaborn==0.11.2
Shapely==1.8.4
six==1.16.0
stack-data==0.5.0
tifffile==2022.8.12
# torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
traitlets==5.3.0
twilio==7.13.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
